Stop asserting that threads are not created during deferred ctors
It's valid for constructors to create threads, so we should allow it.
A potentially problematic case (already handled by locking the deferred
constructor list) is where the constructor calls dlopen and triggers
further constructor calls, so add test coverage for it, The new test is
just a copy of filc/tests/concurrentloadstress with the body of main()
moved to a constructor.